Finally fixed the x64 link issue but it still doesn't load. Added error message detection.
git-svn-id: svn://svn.code.sf.net/p/opticks-extras/code/branches/future/EXTRAS-1/IDL@22 2d91f92e-8166-0410-902e-d52292f61e86

+++ Code/IdlStart/MetadataCommands.cpp
@@ -39,7 +39,7 @@ namespace
       else
       {
          std::vector<T> tmpVal;
-         for (int i = 0; i < total; ++i)
+         for (size_t i = 0; i < total; ++i)
          {
             tmpVal.push_back(pValReal[i]);
          }
@@ -58,7 +58,7 @@ namespace
       else
       {
          std::vector<bool> tmpVal;
-         for (int i = 0; i < total; ++i)
+         for (size_t i = 0; i < total; ++i)
          {
             tmpVal.push_back((pValue[i] == 0) ? false : true);
          }
